backend/flask_services/requirements.txt
flask>=3.0
flask-socketio>=5.3
flask-cors>=4.0
python-dotenv>=1.0
google-generativeai>=0.7
requests>=2.31
reportlab>=4.1
PyGithub>=2.3
gunicorn>=22.0
